+++ app/assets/javascripts/Components/Result/text_viewer.jsx
@@ -31,19 +31,22 @@ export class TextViewer extends React.Component {
     // Remove existing syntax highlighted code.
     $('.dp-highlighter').remove();
     dp.SyntaxHighlighter.HighlightAll(source_id);
-    window.annotation_type = ANNOTATION_TYPES.CODE;
     window.syntax_highlighter_adapter = new SyntaxHighlighter1p5Adapter($('.dp-highlighter ol')[0]);
-
-    window.annotation_manager = new SourceCodeLineAnnotations(
-      new SourceCodeLineManager(
-        window.syntax_highlighter_adapter,
-        new SourceCodeLineFactory(),
-        new SourceCodeLineArray()),
-      new AnnotationTextManager(),
-      new AnnotationTextDisplayer());
 
     // Apply modifications to Syntax Highlighter
     window.syntax_highlighter_adapter.applyMods();
+
+    if (this.props.resultView) {
+      window.annotation_type = ANNOTATION_TYPES.CODE;
+
+      window.annotation_manager = new SourceCodeLineAnnotations(
+        new SourceCodeLineManager(
+          window.syntax_highlighter_adapter,
+          new SourceCodeLineFactory(),
+          new SourceCodeLineArray()),
+        new AnnotationTextManager(),
+        new AnnotationTextDisplayer());
+    }
   };
 
   display_annotation = (annotation) => {